create BudgetStatementComment schema and its read queries
